Use long as the type of aggregate/event version

+++ src/Common/NCafe.Abstractions/Domain/AggregateRoot.cs
@@ -5,7 +5,7 @@ namespace NCafe.Abstractions.Domain;
 public abstract class AggregateRoot
 {
     public Guid Id { get; protected set; }
-    public int Version { get; protected set; } = -1;
+    public long Version { get; protected set; } = -1;
 
     private readonly List<IEvent> _pendingEvents = new();
